Use open class instead of refinements
