+++ lib/bitcoin/store/chain_entry.rb
@@ -3,7 +3,6 @@ module Bitcoin
 
     # wrap a block header object with extra data.
     class ChainEntry
-      using Refinements::EvenLengthHex
 
       attr_reader :header
       attr_reader :height
