--- lib/refinements.rb
@@ -1,1 +1,0 @@
-require_relative 'refinements/event_length_hex'
